fix: make get cfn output more flexible
