+++ frontend/package.json
@@ -9,9 +9,9 @@
     "preview": "vite preview",
     "deploy": "npm run sync && npm run invalidate-cache && npm run print-website-url",
     "publish": "npm run build && npm run deploy",
-    "sync": "aws s3 sync dist/ s3://$(aws cloudformation describe-stacks --output text --query 'Stacks[?contains(StackName, `PetTracker`)][].Outputs[?contains(OutputKey, `StaticAssetsBucketName`)].OutputValue')/ --delete",
-    "print-website-url": "echo \"Website published at $(aws cloudformation describe-stacks --output text --query 'Stacks[?contains(StackName, `PetTracker`)][].Outputs[?contains(OutputKey, `cloudfrontDomainName`)].OutputValue')\n\"",
-    "invalidate-cache": "aws cloudfront create-invalidation --output yaml --distribution-id $(aws cloudformation describe-stacks --output text --query 'Stacks[?contains(StackName, `PetTracker`)][].Outputs[?contains(OutputKey, `cloudfrontDistributionId`)].OutputValue') --paths '/index.html' --no-cli-pager"
+    "sync": "aws s3 sync dist/ s3://$(npm run getCfnOutput -s -w scripts StaticAssetsBucketName)/ --delete",
+    "print-website-url": "echo \"Website published at $(npm run getCfnOutput -s -w scripts cloudfrontDomainName)\n\"",
+    "invalidate-cache": "aws cloudfront create-invalidation --output yaml --distribution-id $(npm run getCfnOutput -s -w scripts cloudfrontDistributionId) --paths '/index.html' --no-cli-pager"
   },
   "dependencies": {
     "@aws-amplify/ui-react": "^5.0.4",
