+++ scripts/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "createConfig": "node create-aws-exports.mjs",
-    "convertCDKtoCfn": "node convert-template.js"
+    "convertCDKtoCfn": "node convert-template.js",
+    "getCfnOutput": "node get-stack-output.mjs"
   },
   "author": {
     "name": "Amazon Web Services",
